+++ src/test/java/kitchenpos/menu/intergration/ProductServiceTest.java
@@ -1,22 +1,23 @@
 package kitchenpos.menu.intergration;
 
+import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.*;
+
+import java.math.BigDecimal;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
 import kitchenpos.common.TestObjectUtils;
 import kitchenpos.integration.IntegrationTest;
 import kitchenpos.menu.command.application.ProductService;
 import kitchenpos.menu.command.domain.product.Product;
 import kitchenpos.menu.command.domain.product.ProductRepository;
 import kitchenpos.menu.infra.JpaProductDao;
-import kitchenpos.menu.query.dto.ProductRequest;
 import kitchenpos.menu.query.dto.ProductResponse;
 import kitchenpos.menu.query.dto.ProductResponses;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import java.math.BigDecimal;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
+import kitchenpos.menu.ui.dto.ProductRequest;
 
 class ProductServiceTest extends IntegrationTest {
 
@@ -30,7 +31,7 @@ class ProductServiceTest extends IntegrationTest {
     @DisplayName("상품을 등록할 수 있다.")
     @Test
     void createTest() {
-        ProductRequest productRequest = new ProductRequest("참치뱃살", 30000);
+        ProductRequest productRequest = new ProductRequest("참치뱃살", BigDecimal.valueOf(30000));
         final ProductResponse product = productService.create(productRequest);
 
         assertAll(
